requirements.txt
Django==5.1.5
psycopg2==2.9.10
python-dotenv==1.0.1
Pillow==11.1.0
django-simple-captcha==0.6.1
django-mathfilters==1.0.0